Separate some classes
